enhancement (core module): more on ontology-cache and ontology service; Knora v2 setup
