+++ projects/knora/core/src/lib/declarations/index.ts
@@ -2,3 +2,5 @@ export * from './core.config';
 export * from './api-service-result';
 export * from './api-service-error';
 export * from './knora-webapi/';
+export * from './utils';
+export * from './knora-constants';
