+++ projects/knora/core/src/lib/services/knora-admin/lists.service.ts
@@ -15,7 +15,7 @@ import {
     ListInfoUpdatePayload,
     ListNodeInfo,
     ListNodeInfoResponse
-} from '../../declarations/index';
+} from '../../declarations';
 import {catchError, map} from 'rxjs/operators';
 
 @Injectable({
